#212 Fix The Bug and Refactor Code of Movement Calculation
The problem was:
The location of the car is stored as X and Y, both being integer. When the movement vector had a very little, but importantly negative value, then the double got always truncated down to the nearest integer, hence the more frame-rate the game has, the faster the car shifts towards negative values when going in certain directions.

Changes:
- took out movement calculation from AutomatedCar
- put move code into its own component
- added support types for the new component for readability